% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Assignment4_Final.R
\name{regress_group_data}
\alias{regress_group_data}
\title{Regression analysis}
\usage{
regress_group_data(data, group_id, obs, covs, include_intercept = TRUE, ...)
}
\arguments{
\item{data}{Data frame that contains all the information}

\item{group_id}{A vector with one or multiple column names that uniquely define the groups}

\item{obs}{A vector with one or multiple column names of targeting observations}

\item{covs}{A vector with one or multiple column names of targeting covariates}

\item{include_intercept}{Interception from regression analysis}

\item{...}{}
}
\value{
Regression analysis of the observations and inputted covariates from using the desired data
}
\description{
Regression analysis
}
