% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Assignment4_Final.R
\name{summarise_group_data}
\alias{summarise_group_data}
\title{Summarize your data frame}
\usage{
summarise_group_data(data, group_id, obs, fun, ...)
}
\arguments{
\item{data}{Data frame that contains all the information}

\item{group_id}{A vector with one or multiple column names that uniquely define the groups.}

\item{obs}{A vector with one or multiple column names of targeting observations}

\item{fun}{A summary function}

\item{...}{Extra arguments for the summary function}
}
\value{
A data fram with group_id columns and all the summary stats
}
\description{
Summarize your data frame
}
\examples{
summarise_group_data(data = iris, group_id = 'Species', obs = c('Petal.Length', 'Sepal.Length'), fun = sd)
}
